resume_tailor/requirements.txt
crewai
crewai[tools]
pandas
python-dotenv
requests
beautifulsoup4
lxml